Reject VMAs without write permission
As it stands now, the IB code always requests writeable regions. So
ensure we don't try to write to a VMA that doesn't have the VM_WRITE
flag set.

I'm not sure if without this it's possible to pass a read-only file to
this code and use RDMA to write to it. I want to ensure this is not
possible.