Improve handling of mmaps invalidation while they are acquired
When the IB core registers these memory regions we take a copy
of the VMA, much the same way forking a process would. If
the user munmaps the memory region before deregistering it,
the IB memory registration is still valid and the backing device can
still maintain a refernce via vm_ops open and close. (This technique is
copied from the v4l2-core in the videobuf2-memops.c file.)

Therefore, the IB invalidation callback is not needed seeing the
memory region is now guaranteed to be valid for the duration of the
registration.