Implementation of io_peer_mem using the mmu_notifier.
This version uses the mmu_notifier API to deal with the potential
security leak issue addressed (badly) in cde507beea0 and our uncertainty
regarding what's required to lock a VM_PFN/IO/MIXEDMAP from changes by
the rest of the kernel.

However, it's my opinion that this approach is wrong due to the fact
that userspace no longer has the guarantee that it's MRs will work
for the duration they are required. Thus I've added a big caveat
as a comment warning about this.