Fix a corner case security leak.
If a malicous userspace process grabs the uverbs file descriptor
(for /dev/infinband/uverbs*), mmaps it, and tries to register
that memory via io_peer_mem then this module will successfully
register the memory, and take a reference to that file.

Seeing the uverbs code only cleans up memory regions only after
that file is closed, and io_peer_mem will only release that file
once the cleanup happens then a deadlock will occur and the
registration will be permanently leaked.

This code handles this situation in a bit of an
hackish way: we check if the backing files, i_cdev's kobjs's name
starts with uverbs and fail if it does.